Fix install path of icons on Linux
The current hack with explicit "install" invocation is broken because it
doesn't respect the $(INSTALL_ROOT) variable. Avoid manual installation
of the icons by naming all of them "chessx.png" and placing them into
individual subdirectories according to the size.

Unfortunately the native Qt install targets do not provide a way to
rename the files when installing, so this is the only way to avoid
running unix commands by hand.

Additionally also fix the destination path of the main icon, which is
now 256 pixels instead of 128.

Signed-off-by: Ondrej Mosnáček <omosnacek@gmail.com>